Fixed bug 1413 - Cannot load images with ImageIO enabled on Mac OS X
The attached diff fixes the library so ImageIO can be used and images are
displayed again without corruption and black screens.

(Based on http://pastebin.com/T3s15UXF and
http://www.brucelindbloom.com/index.html?Eqn_RGB_XYZ_Matrix.html)

I suspect the matrix passed to CGColorSpaceCreateCalibratedRGB() as implemented
in r292 is invalid and the CoreGraphics framework on 10.5 doesn't reject it,
but later ones do. Unfortunately all the documentation I can find doesn't
actually specify what the valid range of values of the arguments is.