Fixed bug 1698 - XPM colors can be in upper case and/or lower case
Sylvain

Due to previous commits, there is a regression : some xpm does not load because
of color name not found.

You called "toupper" before, and now only using "SDL_strncmp" to look for xpm
color name.

In a xpm file, colors can be in Lower and/or Upper case !
So, we need to call "SDL_strncasecmp".