Fixed bug 1991 - XCF and LBM image loading might lead to random crashes
Marcus von Appen

The current XCF and LBM image loaders mix SDL's and the underlying C memory APIs to allocate, reallocate or compare memory, which can lead to random crashes on the target system.

Attached is a small patch to clean up the API and fix a memory lead in the XCF loader implementation.